app/requirements.txt
Django
djangorestframework
django-filter
markdown
django-cors-headers
PyJWT
docxtpl
requests